Add files via upload
The dataset used for this project is the Wine Quality dataset. It contains various chemical properties of wines, such as acidity, pH, alcohol content, etc., along with their corresponding quality ratings. The task is to predict the quality rating of wines based on these features. Make sure to preprocess and clean the dataset before using it for modeling.